Simplify the \Nova\ORM structure
